move type inference to veri engine
