--- cranelift/isle/veri/veri_type_inference/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "veri_type_inference"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-cranelift-isle = { path = "../../isle" }
-veri_annotation = { path = "../veri_annotation" }
-veri_ir = { path = "../veri_ir" }
